move common symbols const and function into separate file (seriesSymbols.js) - also, reduce padding on y_scatter
