+++ packages/perspective-viewer-d3fc/src/js/series/categoryPointSeries.js
@@ -6,12 +6,10 @@
  * the Apache License 2.0.  The full license can be found in the LICENSE file.
  *
  */
-import * as d3 from "d3";
 import * as fc from "d3fc";
 import {tooltip} from "../tooltip/tooltip";
 import {withoutOpacity} from "./seriesColours";
-
-const symbols = [d3.symbolCircle, d3.symbolCross, d3.symbolDiamond, d3.symbolSquare, d3.symbolStar, d3.symbolTriangle, d3.symbolWye];
+import {fromDomain} from "./seriesSymbols";
 
 export function categoryPointSeries(settings, seriesKey, colour, symbols) {
     let series = fc.seriesSvgPoint().size(100);
@@ -34,13 +32,4 @@ export function symbolType(settings) {
     const col = settings.data && settings.data.length > 0 ? settings.data[0] : {};
     const domain = Object.keys(col).filter(k => k !== "__ROW_PATH__");
     return fromDomain(domain);
-}
-
-function fromDomain(domain) {
-    return domain.length > 1
-        ? d3
-              .scaleOrdinal()
-              .domain(domain)
-              .range(symbols)
-        : null;
 }
